Add files via upload
This project is a simple implementation of the classic Tic Tac Toe game. It allows two players (or a player vs
computer, depending on version) to play the game in the terminal/console. The program checks the game board
after every move to determine if there is a winner, a draw, or if the game should continue.
Tools Used: Python